+++ experiments/eval_nni.py
@@ -1,23 +1,19 @@
-from src import update, config
+from src import config
 from elv_client_py import ElvClient
 import tempfile
 import argparse
 import pandas as pd
-import shutil
 import os
 import logging
 import nni
 
-from src.index import FaissIndex
+from src.index.faiss import FaissIndex
 from src.embedding import VideoTagEncoder
-from src.search import SimpleSearcher
-from src.rank import SimpleRanker
-from src.query_understanding import SimpleQueryProcessor
-from src.format import SearchArgs
+from src.ranking.rank import SimpleRanker
+from src.query_processing.simple import SimpleQueryProcessor
+from src.update import builder
 from src.utils import LRUSearchCache
-from experiments.utils import get_score, convert
 from experiments.run_experiment import Experiment, get_experiment_runner
-from src.utils import timeit
 
 def main():
     client = LRUSearchCache(ElvClient.from_configuration_url(config.CONFIG_URL, args.auth))
@@ -30,7 +26,7 @@ def main():
     processor = SimpleQueryProcessor(client, encoder)
     index = FaissIndex(tmp_path, config.IndexConstructor)
     ranker = SimpleRanker(index)
-    index_builder = update.IndexBuilder(encoder)
+    index_builder = builder.IndexBuilder(encoder)
     qid = args.qid
     data = pd.read_csv(os.path.join(config.SRC_PATH, '..', args.data))
     experiment = Experiment(index, index_builder, qid, client, processor, ranker, data, lambda params: encoder.set_t_k(params["T"], params["K"]))
