+++ experiments/optimize_git_clean.py
@@ -1,4 +1,4 @@
-from src import update, config
+from src import config
 from elv_client_py import ElvClient
 import tempfile
 from hyperopt import fmin, tpe, hp, Trials
@@ -8,12 +8,13 @@ import shutil
 import os
 import logging
 
-from src.index import FaissIndex
+from src.index.faiss import FaissIndex
 from src.embedding import VideoTagEncoder
-from src.search import SimpleSearcher
-from src.rank import SimpleRanker
-from src.query_understanding import SimpleQueryProcessor
+from src.search.simple import SimpleSearcher
+from src.ranking.rank import SimpleRanker
+from src.query_processing.simple import SimpleQueryProcessor
 from src.format import SearchArgs
+from src.update import builder
 from src.utils import LRUSearchCache
 from experiments.utils import get_score, convert
 from src.utils import timeit
@@ -32,7 +33,7 @@ def main():
     processor = SimpleQueryProcessor(client, encoder)
     index = FaissIndex(tmp_path, config.IndexConstructor)
     ranker = SimpleRanker(index)
-    index_builder = update.IndexBuilder(encoder)
+    index_builder = builder.IndexBuilder(encoder)
     qid = args.qid
     data = pd.read_csv(args.data)
     experiment = Experiment(index, index_builder, qid, client, processor, ranker, data, lambda params: encoder.set_t_k(params["T"], int(params["K"])))
@@ -47,7 +48,7 @@ def main():
     trials = Trials()
     fmin(fn=run_experiment, space=space, algo=tpe.suggest, max_evals=args.samples, trials=trials)
     # Accessing the results
-    for trial in trials.trials:Ac
+    for trial in trials.trials:
         print(trial['result'], trial['misc']['vals'])
 
 if __name__ == "__main__":
