+++ src/query_processing/simple.py
@@ -1,24 +1,14 @@
 from elv_client_py import ElvClient
 from typing import Dict, Any
-from abc import ABC, abstractmethod
+
 from marshmallow import Schema, fields as field
 import logging
 
-from src.embedding import TextEncoder
+from src.embedding.abstract import TextEncoder
 from src.format import SearchArgs, Any
+from src.query_processing.abstract import QueryProcessor
 
 FieldWeights = Dict[str, float]
-
-# Abstract class representing Query Understanding block
-class QueryProcessor(ABC):
-    # Args:
-    #   content_id: id of the index object
-    #   query: user query
-    #  Returns:
-    #   A processed query, the type of which is defined by the subclass. Will be passed to core search block.
-    @abstractmethod
-    def process_query(self, content_id: str, query: SearchArgs) -> Any:
-        pass
 
 class SimpleQueryProcessor(QueryProcessor):
