+++ src/ranking/rank.py
@@ -1,26 +1,16 @@
-from typing import List, Any
+
 import numpy as np
 from typing import List, Dict, Callable, Tuple, Iterable
 import numpy as np
 from sentence_transformers import util as ut
-from abc import ABC, abstractmethod
 
-from src.index import Index, VectorDocument
-from src.query_understanding import SimpleQueryProcessor
+from src.index.abstract import Index, VectorDocument
+from src.query_processing.simple import SimpleQueryProcessor
+
+from src.ranking.abstract import Ranker
 
 # function taking (query_embedding, VectorDocument) -> score
 Scorer = Callable[[np.ndarray, VectorDocument], float] 
-
-class Ranker(ABC):
-    """
-    Abstract class representing the Re-ranking block in the design. This class is meant to be subclassed by a specific ranking implementation. 
-    The rank method should take a list of uids and a query and return a list of uids and their associated scores (in sorted order).
-
-    The query format is undefined here, the specific implementation will define the expected input (see SimpleRanker example below)
-    """
-    @abstractmethod
-    def rank(self, uids: Iterable[str], limit: int, query: Any) -> List[Tuple[str, float]]:
-        pass
 
 class SimpleRanker(Ranker):
     """
